requirements.txt
numpy>=1.19.4
pandas>=1.1.4
scipy>=1.5.3
matplotlib>=3.3.3
gym>=0.17.3
progressbar2>=3.53.1
pymdptoolbox>=4.0-b3
